FOR FUTURE USE
This is a pretty important update for future use and for developing POST request parsing technique.
We're going to communicate with the API using ArrayList<Byte>, which allows the application to contain non-string characters and store binary and non-printable characters.
This is helpful and it's a requirement for POST binary file upload to work probably.
Sooner or later, the development team "It's me only tbh" will implement POST data Parsing in another global ArrayList<Byte> so you can deal with it!
The length of the POST data variable will be max. size of 1MB "Increase it at your own risk".
Also, CPU and Memory improvements are coming soon.